:recycle: [Refactor] useDetailDialog 폴더 위치 변경 #main
