+++ src/features/ui/user/UserDetailDialog.tsx
@@ -1,5 +1,5 @@
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../../../shared/ui"
-import { usePost } from "../../../features/model/post/usePost"
+import { usePost } from "../../model/post/usePost"
 
 export const UserDetailDialog = () => {
   const { showUserModal, setShowUserModal, selectedUser: user } = usePost()
